requirements.txt
Flask==3.1.2
Flask-SocketIO==5.6.0
gunicorn==25.1.0
eventlet==0.40.4
psycopg2-binary==2.9.11
pandas==2.3.3